Use 'slim' structures
LSG now uses the `-end-pos` file produced by BEETL (instead of reconstructing the permutation with a scan of the BWT and of the GSA). Moreover, it now reads a SA without the sequence IDs (but only the suffix lengths) and sequence lengths (either of the LCP or of the GSA) are assumed to be 8-bit long.
Requires a customized version of BEETL (found in directory `thirdparty`).